Fixed a segfault in rs_buf_str().
It was triggered by the input "#b" (or any other radix letter).
ST_BINARY (for example) would call check_num with an empty buffer, which
in turn called rs_buf_str(). rs_buf_str() writes a '\0' character to the
buffer in order to terminate the string. Since the buffer was empty
(buf->buf == NULL), trying to write to it caused a segfault. Two changes
were made:

* rs_buf_str() returns an empty string if the buffer is empty. This
  prevents the segfault.
* ST_HASH explicitly checks to see if a radix character is followed by a
  delimiter. This prevents check_num() from being called with an empty
  buffer in the first place. (Otherwise, "#b" would evaluate to 0).

The moral of this story: Input isn't always reasonable, so don't just
test reasonable input.